Dev copernicus

+++ 5_model_input/SUMMA/3f_attributes/2b_insert_landclass_from_hist_into_attributes.py
@@ -97,7 +97,7 @@ with nc4.Dataset(attribute_path/attribute_name, "r+") as att:
         attribute_hru = att['hruId'][idx]
     
         # Find the row in the shapefile that contains info for this HRU
-        shp_mask = (shp[intersect_hruId_var] == attribute_hru)
+        shp_mask = (shp[intersect_hruId_var].astype(int) == attribute_hru)
         
         # Extract the histogram values
         tmp_hist = []
